use mongooseimctl instead reusing xmpp_session
When reusing the xmpp_session, we cannot set the presence of xmpp client
directly connected to the server. Same thing for kicking xmpp client
directly connected to the server.

Morever, the get_presence will use the command line to retrieve the
presence.

The best should be have admin HTTP REST API instead of command line, but
for now we just use the command line. It shouldn't be to complicate to
pay for REST API mapping command line.

For now, we update the presence of every resource of a user
whatever its priority. So when the presence "disconnected" is set, every
resource is kicked out.

TODO:
* delete the xmpp_session_collections
* The user "xivo-websocketd" need to have permission to execute
  "mongooseimctl"
* acceptance tests