fix: validate the correct token
The period task was a bit bugged it was retrieving the token from the
websocket session, wait one minute, then validate the token.

So if new token is sent by the client to the websocket session, we end
up to not validate the correct one.

This ensure we always validate the latest one.