bus: bind on wazo-headers instead of wazo-websocketd
Why:

* wazo-websocketd exchange is no longer needed since
it doesn't provide anything different than wazo-headers.

The exchange is legacy and was used when bound on xivo(topic)
by unbinding AMI events since we never want to send them through
the websocket.  With the new filtering rules, these events are
dropped directly in RabbitMQ and will never reach the endusers